convert to FC, break out helpers
